fix: moved out components fro pages fix: removed next export to support client side render
