+++ app/services/root.js
@@ -8,9 +8,9 @@ export const getReccomendations = async () => {
       console.error(res.data);
       return [];
     }
-    // return response.data.map(({ id, name }) => ({ id, name }));
-    const desired = ['react-floki', 'nextjs-template'];
-    return response.data.filter(({ name }) => desired.includes(name)).map(({ id, name }) => ({ id, name }));
+
+    const recommendations = ['react-floki', 'nextjs-template'];
+    return response.data.filter(({ name }) => recommendations.includes(name)).map(({ id, name }) => ({ id, name }));
   };
   return getData(res);
 };
