+++ package.json
@@ -3,15 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "build:prod": "env-cmd -f environments/.env.production next build && next export",
-    "build:qa": "env-cmd -f environments/.env.qa next build && next export",
-    "build:develop": "env-cmd -f environments/.env.development next build && next export",
+    "build:prod": "env-cmd -f environments/.env.production next build",
+    "build:qa": "env-cmd -f environments/.env.qa next build",
+    "build:develop": "env-cmd -f environments/.env.development next build",
     "start:prod": "env-cmd -f environments/.env.production next start",
     "start:qa": "env-cmd -f environments/.env.qa next dev",
     "start:dev": "env-cmd -f environments/.env.development next dev",
     "dev": "next dev",
     "custom:dev": "env-cmd -f environments/.env.development node server/index.js",
-    "build": "next build && next export",
+    "build": "next build",
     "start": "next start",
     "test": "jest --coverage",
     "test:ci": "jest --ci",
