--- pages/_root/index.js
@@ -1,5 +1,0 @@
-export { default as types } from './typedef';
-export { default as RepoList } from './RepoList';
-export { default as ErrorState } from './ErrorState';
-export { default as Recommended } from './Recommended';
-export { default as StyledComponents } from './styled';
