--- pages/_root/typedef.js
@@ -1,18 +1,0 @@
-import PropTypes from 'prop-types';
-
-const propTypes = {
-  reposData: PropTypes.arrayOf(
-    PropTypes.shape({
-      totalCount: PropTypes.number,
-      incompleteResults: PropTypes.bool,
-      items: PropTypes.array
-    })
-  ),
-  reposError: PropTypes.object,
-  repoName: PropTypes.string,
-  recommendations: PropTypes.arrayOf(
-    PropTypes.shape({ id: PropTypes.number.isRequired, name: PropTypes.string.isRequired })
-  )
-};
-
-export default propTypes;
