--- pages/info/_name/index.js
@@ -1,2 +1,0 @@
-export { default as types } from './typedef';
-export { default as Title } from './Title';
