--- pages/info/_name/typedef.js
@@ -1,8 +1,0 @@
-import PropTypes from 'prop-types';
-
-const propTypes = {
-  name: PropTypes.string.isRequired,
-  stargazersCount: PropTypes.number.isRequired
-};
-
-export default propTypes;
